#266: Split analyzer into analyzer and analyzer-explain
